Job submissions can cut on recoil-pion mass
The combined recoil proton and pion mass are
typically unrelated, but can peak when both
particles are the result of an excited baryon
decay like the Delta. Submitter now has option
to cut on this mass spectrum, just like the TEM
cuts, to remove these baryon contributions.

Directory paths have been updated to accommodate
this change