Major revision to data getter, and new script
get_bin_data now does not perform any cuts, but
simply obtains the parameters of interest from
the pre-cut ROOT trees that were made in the
submission process. Also added columns for the
-t momentum transfer RMS and MEAN.

A new script to aggregate all bootstrap fits
for a particular fit was added. Is very slow and
needs speed improvements

Other changes:
- Adjusted pwa_tools for new data getter method