Initial commit to develop a new config writer
The current config writer works using built-in
AmpTools loops. These are more readable, but do
not offer a simple way to remove single waves
of a particular reflectivity. This branch
will switch the config writer to instead write
out each amplitude individually in a loop,
effectively mirroring how AmpTools expands the
LOOP functions into all individual amplitudes.

This is a major rewrite of the config writer and
so will be done in a separate branch.