Separated data and phasespace file name options
The submission process assumed a more rigid file
naming scheme for Monte Carlo file versions and
their selector options. This commit separates
the data and phasespace version, directory, and
selector option into 3 cli args each. This avoids
errors when the user wants to specificy a
phasespace selector option, that previously would
be included with the "version" variable and cause
an error for the Gen trees.

This commit also includes:
- Small README update
- Function for improved consistency submit.py