Phase constraint handles missing reflectivities
The phase constraint argument assumed that both
reflectivities were present for a JpmL wave. This
is not always the case, especially when removing
waves. This commit changes the behavior of the
argument to now explicitly ask for both reference
waves, and if not set, will only successfully run
as long as a wave with both reflectivities is
present.

This commit also adds:
- A small fix for a constant Path variable
- Change arguments to use "word-word" format