Moment projection now uses .fit files
The .txt files have a bug where parameters
included from another file are not saved to the
output "seed" .txt file. This was causing an issue
where the scale of the production parameters
was not being recorded and used. The .fit files
do not have this bug, and still contain all the
relevant information.