WIP: Have moment projection produce a matrix csv
This commit is partially done work on getting the
moment projection script to also produce a csv,
that is a matrix of the moments and production
coefficients that make up those moments

The script is functioning and produces a csv, but
has far less production coefficient pairs than
expected. The moment values were verified not to
have changed, but the clebsch gordans (entries of
the csv) have not yet been verified.