Overhaul of angle plotter for automated plotting
The clunky maps were abandoned in favor of a
constant property map for the jp values, that
mimicked the plotting behavior of pwa_tools.py.
This alignment also necessitated changing some of
the python markers to match those available in
ROOT.

It also includes automatic handling of what jp
values are present, avoiding a previous error.