Proper env file creation, and updated c++ debugger
The launch.json file was not correctly supporting
debug flags due to the binary being a child
process of bash. The binary is now directly
launched with a corresponding env file, with a
caveat that a variable number of arguments is not
supported yet by the cpp extension. For now we'll
need to hardcode launch configs for each # of
arguments.