added symbolic links to root logon scripts
these scripts live in the home directory, but I've
had issues with it getting wiped in the past, and
these are highly specialized and imoprtant for any
interactive ROOT sessions (like done in FSRoot).

Since I don't want to git track my entire home
directory, I'm adding symlinks to these files in
the neutralb1 repo, so they can be easily restored
if needed.