Make bkg cubic and explicitly set omega pi masses
Linear fit was not good enough for the background
under the omega peak, so a cubic is now used.
Also, the omega pi0 mass bin center is explicitly
saved in the different containers for easier
tracking of fit parameters.

This commit also includes saving / loading of
histogram caches, and renaming of some variables
for readability.